fix: invalidate cache on files imported from tests using replace
If a file is imported from a file that has done a replace, the file will
now be requested with a unique url. This prevents a bug where files that
had already been imported from somewhere, would not receive replaced
versions of sub dependencies. This will only work if the file is
required from the same test file as the replacement is made from
however. It is not a perfect solution but should be the common case.