fix: make reset work
This makes reset work properly. After a reset, no special urls should be
generated by resolve. The counts still needs to be kept since if a
replaced file is reset and replaced again, the count should not be the
same as with the first replace.

To solve this, arrays listing the urls of files currently under
replacement was added to the info. If a file is not in this array, no
count will be generated.